--- src/main/java/me/xhsun/guildwars2wrapper/requests/v2/commerce/Prices.java
@@ -1,61 +1,0 @@
-package me.xhsun.guildwars2wrapper.requests.v2.commerce;
-
-import java.util.List;
-import me.xhsun.guildwars2wrapper.error.GuildWars2Exception;
-import me.xhsun.guildwars2wrapper.model.v2.commerce.Listing;
-import me.xhsun.guildwars2wrapper.requests.GuildWars2API;
-import me.xhsun.guildwars2wrapper.requests.Request;
-import retrofit2.Call;
-import retrofit2.Callback;
-import retrofit2.Response;
-
-public class Prices extends Request {
-
-	protected Prices(GuildWars2API gw2API) {
-		super(gw2API);
-	}
-
-	//ASYNC
-	/**
-	 * For more info on Listing Price API go <a href="https://wiki.guildwars2.com/wiki/API:2/commerce/prices">here</a><br/>
-	 * Give user the access to {@link Callback#onResponse(Call, Response)} and {@link Callback#onFailure(Call, Throwable)} methods for custom interactions
-	 *
-	 * @param ids      list of item id
-	 * @param callback callback that is going to be used for {@link Call#enqueue(Callback)}
-	 * @throws GuildWars2Exception  empty ID list
-	 * @throws NullPointerException if given {@link Callback} is empty
-	 * @see Listing listing item price info
-	 */
-	public void getTPListingInfo(int[] ids, Callback<List<Listing>> callback) throws GuildWars2Exception, NullPointerException {
-		isParamValid(new ParamChecker(ids));
-		gw2API.getTPListingInfo(processIds(ids)).enqueue(callback);
-	}
-
-	/**
-	 * For more info on Listing Price API go <a href="https://wiki.guildwars2.com/wiki/API:2/commerce/prices">here</a><br/>
-	 * Give user the access to {@link Callback#onResponse(Call, Response)} and {@link Callback#onFailure(Call, Throwable)} methods for custom interactions
-	 *
-	 * @param callback callback that is going to be used for {@link Call#enqueue(Callback)}
-	 * @throws NullPointerException if given {@link Callback} is empty
-	 */
-	public void getAllPriceID(Callback<List<Integer>> callback) throws NullPointerException {
-		gw2API.getAllTPPriceIDs().enqueue(callback);
-	}
-
-	/**
-	 * For more info on Listing Price API go <a href="https://wiki.guildwars2.com/wiki/API:2/commerce/prices">here</a><br/>
-	 * Give user the access to {@link Callback#onResponse(Call, Response)} and {@link Callback#onFailure(Call, Throwable)} methods for custom interactions
-	 *
-	 * @param ids      list of item id
-	 * @param callback callback that is going to be used for {@link Call#enqueue(Callback)}
-	 * @throws GuildWars2Exception  empty ID list
-	 * @throws NullPointerException if given {@link Callback} is empty
-	 * @see me.xhsun.guildwars2wrapper.model.v2.commerce.Prices listing item price info
-	 */
-	public void getPriceInfo(int[] ids, Callback<List<me.xhsun.guildwars2wrapper.model.v2.commerce.Prices>> callback) throws GuildWars2Exception, NullPointerException {
-		isParamValid(new ParamChecker(ids));
-		gw2API.getTPPriceInfo(processIds(ids)).enqueue(callback);
-	}
-
-	//SYNC
-}
